+++ app/Http/Livewire/Propostas/DetalheProposta.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Livewire\Propostas;
 
+use Livewire\Component;
+use App\Models\Carrinho;
+use Livewire\WithPagination;
+use Illuminate\Support\Facades\Auth;
 use App\Interfaces\ClientesInterface;
 use App\Interfaces\PropostasInterface;
-use App\Models\Carrinho;
 use Illuminate\Support\Facades\Session;
-use Livewire\Component;
-use Livewire\WithPagination;
 
 class DetalheProposta extends Component
 {
@@ -232,7 +233,7 @@ class DetalheProposta extends Component
     public function deletartodos()
     {
         $this->detailsClientes = $this->clientesRepository->getDetalhesCliente($this->idCliente);
-        Carrinho::where('id_cliente', $this->detailsClientes->customers[0]->no)->delete();
+        Carrinho::where('id_cliente', $this->detailsClientes->customers[0]->no)->where('id_user',Auth::user()->id)->delete();
 
         $this->dispatchBrowserEvent('itemDeletar');
     }
@@ -540,6 +541,7 @@ class DetalheProposta extends Component
         }
 
         $this->carrinhoCompras = Carrinho::where('id_cliente', $this->detailsClientes->customers[0]->no)
+            ->where('id_user',Auth::user()->id)
             ->where('id_proposta', '!=', '')
             ->get();
